fix: settingIcon을 UnifiedSettingIcon으로 변경 및 관련 아이콘 수정
