+++ src/assets/svg/LargeClose.tsx
@@ -1,4 +1,4 @@
-import { SvgProps } from './nav/SettingIcon';
+import { SvgProps } from './nav/UnifiedSettingIcon';
 
 export const LargeClose = ({ className, ...props }: SvgProps) => {
   return (
