+++ src/assets/svg/SparkIcon.tsx
@@ -1,4 +1,4 @@
-import { SvgProps } from './nav/SettingIcon';
+import { SvgProps } from './nav/UnifiedSettingIcon';
 
 export const SparkIcon = ({ className }: SvgProps) => {
   return (
