+++ src/assets/svg/logo/InstaGram.tsx
@@ -1,4 +1,4 @@
-import { SvgProps } from '../nav/SettingIcon';
+import { SvgProps } from '../nav/UnifiedSettingIcon';
 
 export const InstaGram = ({ ...props }: SvgProps) => {
   return (
