+++ src/assets/svg/logo/Logo.tsx
@@ -1,4 +1,4 @@
-import { SvgProps } from '../nav/SettingIcon';
+import { SvgProps } from '../nav/UnifiedSettingIcon';
 
 export const Logo = ({ className, ...props }: SvgProps) => {
   return (
