+++ src/assets/svg/logo/SmallYoutbeIcon.tsx
@@ -1,0 +1,25 @@
+import { SvgProps } from '../nav/UnifiedSettingIcon';
+
+export const SmallYoutubeIcon = ({ ...props }: SvgProps) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M16.8922 6.1617C16.842 5.97422 16.6882 5.82045 16.5008 5.77031C16.0655 5.65396 15.5889 5.63968 15.1424 5.60693C14.4618 5.55702 13.6655 5.52296 12.8967 5.49999C12.1306 5.47711 11.4052 5.46562 10.8704 5.45987C10.6034 5.457 10.3847 5.45557 10.2332 5.45485C10.1575 5.45449 10.0986 5.45432 10.0589 5.45423H9.94107C9.90141 5.45432 9.84253 5.45449 9.76681 5.45485C9.61534 5.45557 9.39664 5.457 9.12959 5.45987C8.59482 5.46562 7.86937 5.47711 7.10326 5.49999C6.33447 5.52296 5.53818 5.55702 4.85765 5.60693C4.41147 5.63965 3.93405 5.65325 3.49923 5.77031C3.31175 5.82045 3.15797 5.97422 3.10783 6.1617C2.90895 6.90553 2.87587 7.70646 2.836 8.47146C2.81386 8.89631 2.80273 9.28166 2.79716 9.56089C2.79438 9.70014 2.79301 9.81206 2.79233 9.88821L2.79167 9.99511L2.79233 10.1117C2.79301 10.1879 2.79438 10.2998 2.79716 10.4391C2.80273 10.7183 2.81386 11.1036 2.836 11.5285C2.88289 12.4282 2.97191 13.3334 3.10711 13.8356C3.1582 14.0253 3.31217 14.1796 3.50191 14.2304C3.67417 14.2767 4.13012 14.3397 4.85765 14.393C5.53818 14.4429 6.33447 14.477 7.10326 14.5C7.86937 14.5228 8.59482 14.5343 9.12959 14.5401C9.39664 14.543 9.61534 14.5444 9.76681 14.5451C9.84253 14.5455 9.90141 14.5456 9.94107 14.5457H10.0589C10.0986 14.5456 10.1575 14.5455 10.2332 14.5451C10.3847 14.5444 10.6034 14.543 10.8704 14.5401C11.4052 14.5343 12.1306 14.5228 12.8967 14.5C13.6655 14.477 14.4618 14.4429 15.1424 14.393C15.8699 14.3397 16.3258 14.2767 16.4981 14.2304C16.6865 14.1797 16.8416 14.0273 16.8922 13.8382C17.091 13.0944 17.1241 12.2935 17.164 11.5285C17.1861 11.1036 17.1973 10.7183 17.2028 10.4391C17.2056 10.2998 17.207 10.1879 17.2077 10.1117L17.2083 10.0004L17.2083 9.97431L17.2077 9.88821C17.207 9.81206 17.2056 9.70014 17.2028 9.56089C17.1973 9.28166 17.1861 8.89631 17.164 8.47146C17.1242 7.70701 17.0923 6.90477 16.8922 6.1617ZM18.875 9.99998C18.875 9.99998 18.875 12.8843 18.5022 14.2688C18.2981 15.0321 17.6946 15.6356 16.9314 15.8397C15.5469 16.2125 10 16.2125 10 16.2125C10 16.2125 4.45312 16.2125 3.06862 15.8397C2.30537 15.6356 1.70187 15.0321 1.49775 14.2688C1.125 12.8843 1.125 9.99998 1.125 9.99998C1.125 9.99998 1.125 7.1156 1.49775 5.7311C1.70187 4.96785 2.30537 4.36435 3.06862 4.16023C4.45312 3.78748 10 3.78748 10 3.78748C10 3.78748 15.5469 3.78748 16.9314 4.16023C17.6946 4.36435 18.2981 4.96785 18.5022 5.7311C18.875 7.1156 18.875 9.99998 18.875 9.99998Z"
+        fill="#5E6166"
+      />
+      <path
+        d="M8.84993 12.302C8.57215 12.4622 8.22504 12.2618 8.22504 11.9411V8.0589C8.22504 7.73821 8.57215 7.53774 8.84993 7.69799L12.2145 9.63907C12.4924 9.79942 12.4924 10.2005 12.2145 10.3609L8.84993 12.302Z"
+        fill="#5E6166"
+      />
+    </svg>
+  );
+};
