+++ src/assets/svg/nav/BackIcon.tsx
@@ -1,4 +1,4 @@
-import { SvgProps } from './SettingIcon';
+import { SvgProps } from './UnifiedSettingIcon';
 
 export const BackIcon = ({
   color = 'black',
