+++ src/assets/svg/nav/CloseIcon.tsx
@@ -1,4 +1,4 @@
-import { SvgProps } from './SettingIcon';
+import { SvgProps } from './UnifiedSettingIcon';
 
 export const CloseIcon = ({
   color = 'black',
