+++ src/assets/svg/nav/HomeIcon.tsx
@@ -1,4 +1,4 @@
-import { SvgProps } from './SettingIcon';
+import { SvgProps } from './UnifiedSettingIcon';
 
 export const HomeIcon = ({ color, className }: SvgProps) => {
   return (
