--- src/assets/svg/nav/SettingIcon.tsx
@@ -1,28 +1,0 @@
-import type { ComponentProps } from 'react';
-
-export interface SvgProps extends ComponentProps<'svg'> {
-  color?: string;
-  className?: string;
-  fill?: string;
-  onClick?: () => void;
-}
-
-export const SettingIcon = ({ color, className }: SvgProps) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="25"
-      height="24"
-      viewBox="0 0 25 24"
-      fill="gray"
-      className={className}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M14.2442 2.71198C13.316 2.17608 12.1724 2.17608 11.2442 2.71198L5.45 6.05728C4.5218 6.59318 3.95 7.58356 3.95 8.65536V15.346C3.95 16.4178 4.5218 17.4081 5.45 17.944L11.2442 21.2893C12.1724 21.8252 13.316 21.8252 14.2442 21.2893L20.0385 17.944C20.9667 17.4081 21.5385 16.4178 21.5385 15.346V8.65536C21.5385 7.58356 20.9667 6.59318 20.0385 6.05728L14.2442 2.71198ZM12.2442 4.44403C12.5536 4.2654 12.9348 4.2654 13.2442 4.44403L19.0385 7.78933C19.3479 7.96797 19.5385 8.29809 19.5385 8.65536V15.346C19.5385 15.7032 19.3479 16.0333 19.0385 16.212L13.2442 19.5573C12.9348 19.7359 12.5536 19.7359 12.2442 19.5573L6.45 16.212C6.1406 16.0333 5.95 15.7032 5.95 15.346V8.65536C5.95 8.29809 6.1406 7.96797 6.45 7.78933L12.2442 4.44403ZM11.2442 12.0007C11.2442 11.1723 11.9158 10.5007 12.7442 10.5007C13.5727 10.5007 14.2442 11.1723 14.2442 12.0007C14.2442 12.8292 13.5727 13.5007 12.7442 13.5007C11.9158 13.5007 11.2442 12.8292 11.2442 12.0007ZM12.7442 8.50073C10.8112 8.50073 9.24423 10.0677 9.24423 12.0007C9.24423 13.9337 10.8112 15.5007 12.7442 15.5007C14.6772 15.5007 16.2442 13.9337 16.2442 12.0007C16.2442 10.0677 14.6772 8.50073 12.7442 8.50073Z"
-        fill={color}
-      />
-    </svg>
-  );
-};
